Hold Upgrade

+++ Assets/Scripts/GameDatabase/OreDatabase.cs
@@ -26,13 +26,13 @@ namespace GameDatabase
                 public string oreName;
                 public Sprite oreSprite;
                 public string oreDescription;
-                public decimal currentHardness;
+                public float currentHardness;
                 public bool isUnlocked;
 
                 public int amount;
     
                 [Header("Default Values")]
-                public decimal defaultHardness;
+                public float defaultHardness;
                 public bool isPremium;
         }
 }
